# swagger.yaml
...
swagger: 2.0
info:
  description: Used in MyRA
  version: 1.0.0
  title: MyRA API
  contact:
    email: bcdevexchange@gov.bc.ca
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /AidanLaing/myra_api/1.0.0
...
schemes:
  - https
paths:
...
  /v1/agreement/{agreementId}/zone:
    put:
      tags:
        - Agreement
      summary: Update an agreement zone
      description: Uses the zoneId from body to update the agreement zone.
      operationId: updateAgreementZone
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: zoneId
          description: Zone id that agreement will be updated to.
          required: true
          schema:
            $ref: '#/definitions/zoneId'
        - name: agreementId
          in: path
          description: Id of parent agreement.
          required: true
          type: string
      responses:
        200:
          description: Zone updated successfully
          schema:
            $ref: '#/definitions/Zone'
        400:
          description: Both agreementId and zoneId must be provided.
        401:
          description: Not authorized.
        404:
          description: Agreement or Zone not found.
      security:
        - bearerAuth: []
...
  /v1/client:
    get:
      tags:
        - Client
      summary: Get all clients
      description: Returns an array of all clients.
      operationId: readAllClients
      produces:
        - application/json
      parameters: []
      responses:
        200:
          description: Successfully returned all clients.
          schema:
            type: array
            items:
              $ref: '#/definitions/Client'
        401:
          description: Not authorized.
      security:
        - bearerAuth: []
  /v1/client/{clientId}:
    get:
      tags:
        - Client
      summary: Returns client matching id
      description: Returns the client matching the associated client id.
      operationId: readClient
      produces:
        - application/json
      parameters:
        - name: clientId
          in: path
          description: Id of client to return.
          required: true
          type: integer
      responses:
        200:
          description: Successfully returned client extension matching id.
          schema:
            $ref: '#/definitions/Client'
        401:
          description: Not authorized.
        404:
          description: Client not found.
      security:
        - bearerAuth: []
...
  /v1/zone:
    get:
      tags:
        - Zones
      summary: Returns zones
      description: Returns an array of zones based on query params. Passing no query params will result in all zones being returned.
      operationId: readZones
      produces:
        - application/json
      parameters:
        - name: districtId
          in: query
          description: Pass an optional district id to get all zones within that district.
          required: false
          type: string
      responses:
        200:
          description: Sucessfully returned zones matching query params.
          schema:
            type: array
            items:
              $ref: '#/definitions/Zone'
        400:
          description: Bad query params passed.
        401:
          description: Not authorized.
      security:
        - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  Zone:
    type: object
    required:
      - code
      - description
      - districtId
    properties:
      id:
        type: integer
        example: 58
      code:
        type: string
        example: DOS3
      description:
        type: string
        example: Lakes TSA South of Francois Lake
      contactName:
        type: string
        example: Rob Jimmy
      contactPhoneNumber:
        type: string
        example: 1234567890
      contactEmail:
        type: string
        example: robjimmy@gov.bc.ca
      districtId:
        type: integer
        example: 124124
      createdAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      updatedAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      district:
        $ref: '#/definitions/District'
    example:
      createdAt: 2018-03-15T16:52:37.658Z
      code: DOS3
      districtId: 124124
      contactEmail: robjimmy@gov.bc.ca
      contactName: Rob Jimmy
      district:
        createdAt: 2018-03-15T16:52:37.658Z
        code: BEAV
        description: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
        id: 15
        updatedAt: 2018-03-15T16:52:37.658Z
      description: Lakes TSA South of Francois Lake
      contactPhoneNumber: 1234567890
      id: 58
      updatedAt: 2018-03-15T16:52:37.658Z
  District:
    type: object
    properties:
      id:
        type: integer
        example: 15
      code:
        type: string
        example: BEAV
      description:
        type: string
        example: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
      createdAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      updatedAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
    example:
      createdAt: 2018-03-15T16:52:37.658Z
      code: BEAV
      description: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
      id: 15
      updatedAt: 2018-03-15T16:52:37.658Z
...
  Client:
    type: object
    required:
      - location
    properties:
      id:
        type: integer
        example: 127412
      location:
        type: string
        example: XX
      name:
        type: string
        example: Eriberto Fisher V
      clientStartDate:
        type: string
        example: 2018-03-15T16:52:37.658Z
      clientEndDate:
        type: string
        example: 2018-03-15T16:52:37.658Z
      createdAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      updatedAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      clientType:
        $ref: '#/definitions/ClientType'
    example:
      createdAt: 2018-03-15T16:52:37.658Z
      clientType:
        createdAt: 2018-03-15T16:52:37.658Z
        code: A
        description: Licensee
        active: true
        id: 127412
        updatedAt: 2018-03-15T16:52:37.658Z
      clientEndDate: 2018-03-15T16:52:37.658Z
      name: Eriberto Fisher V
      location: XX
      id: 127412
      clientStartDate: 2018-03-15T16:52:37.658Z
      updatedAt: 2018-03-15T16:52:37.658Z
  ClientType:
    type: object
    required:
      - active
      - code
      - description
    properties:
      id:
        type: integer
        example: 127412
      code:
        type: string
        example: A
      description:
        type: string
        example: Licensee
      active:
        type: boolean
        example: true
      createdAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      updatedAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
    example:
      createdAt: 2018-03-15T16:52:37.658Z
      code: A
      description: Licensee
      active: true
      id: 127412
      updatedAt: 2018-03-15T16:52:37.658Z
...
  zoneId:
    type: object
    properties:
      zoneId:
        type: integer
        example: 1